+++ src/main/java/com/mdh/devtable/waiting/infra/persistence/WaitingRepository.java
@@ -1,6 +1,6 @@
 package com.mdh.devtable.waiting.infra.persistence;
 
-import com.mdh.devtable.ownerwaiting.presentaion.dto.WaitingInfoResponseForOwner;
+import com.mdh.devtable.ownerwaiting.application.dto.WaitingInfoResponseForOwner;
 import com.mdh.devtable.waiting.domain.Waiting;
 import com.mdh.devtable.waiting.domain.WaitingStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
